Change SPI_BCM2835, SPI_BCM2835AUX from y to m
WHY: poor detection of USB of EC21-J, and even if it could be detected, it is not enabled (no LED lights up and no AT command response)

Changing the driver to kernel module fixed above.
The EC-21J worked fine on the Raspberry Pi OS, so I found this fix by
comparing nerves config with the linux config and doing bisection search. I do not understand why this fix it.